:pencil_2: boilerplate to setup navigation for mylist
