+++ src/Home/MyList/MyListView/index.ts
@@ -1,0 +1,1 @@
+export { default as MyList } from "./MyListView";
